Add command reference docs
