+++ glide.lock
@@ -1,5 +1,5 @@
-hash: faa3615182a871c638ea276ae3d4b2d3b42e076e3428e737e596cd0e1e9ac2f1
-updated: 2017-06-20T05:52:27.037860803-07:00
+hash: 4603d45081832651b419c72d01373806592fe66c295a144567aab3fdcbedcda8
+updated: 2017-06-27T07:17:40.856435772-07:00
 imports:
 - name: cloud.google.com/go
   version: fe3d41e1ecb2ce36ad3a979037c9b9a2b726226f
@@ -11,7 +11,7 @@ imports:
   subpackages:
   - h/log
 - name: github.com/appscode/go
-  version: 5eabd32a8a0fe824b5e12c37dc53a0afea8507c9
+  version: ecda61fa4e8943020e9948723e49c4ae544f7a3a
   subpackages:
   - arrays
   - crypto/rand
@@ -86,6 +86,10 @@ imports:
   version: 0d446f3052e47b5ad95989d9a8ffe4f6675e69b5
   subpackages:
   - pkg/client/monitoring/v1alpha1
+- name: github.com/cpuguy83/go-md2man
+  version: 71acacd42f85e5e82f70a55327789582a5200a90
+  subpackages:
+  - md2man
 - name: github.com/davecgh/go-spew
   version: 5215b55f46b2b919f50a1df0eaa5886afe4e3b3d
   subpackages:
@@ -133,7 +137,7 @@ imports:
 - name: github.com/facebookgo/stack
   version: d9d227395bc3d7a635388c75b2e3b3a750e11888
 - name: github.com/flosch/pongo2
-  version: 1d0f0d3af150c4a65dfd424d742f7374819e7d29
+  version: 9d978217a2f84fab35dd17731da1e1c23be32d14
 - name: github.com/ghodss/yaml
   version: 73d445a93680fa1a78ae23a5839bad48f32ba1ee
 - name: github.com/go-ini/ini
@@ -178,7 +182,7 @@ imports:
 - name: github.com/jpillora/go-ogle-analytics
   version: 14b04e0594ef6a9fd943363b135656f0ec8c9d0e
 - name: github.com/juju/errors
-  version: a3edbf48b75d756b79f3abb22bfc8a56e349081f
+  version: 6bffea730f4aba190658f565cb02e95679f7ce31
 - name: github.com/juju/ratelimit
   version: 77ed1c8a01217656d2080ad51981f6e99adaa177
 - name: github.com/lib/pq
@@ -226,10 +230,16 @@ imports:
   version: 8a290539e2e8629dbc4e6bad948158f790ec31f4
 - name: github.com/PuerkitoBio/urlesc
   version: 5bd2802263f21d8788851d5305584c82a5c75d7e
+- name: github.com/russross/blackfriday
+  version: 300106c228d52c8941d4b3de6054a6062a86dda3
+- name: github.com/shurcooL/sanitized_anchor_name
+  version: 10ef21a441db47d8b13ebcc5fd2310f636973c77
 - name: github.com/sirupsen/logrus
   version: ba1b36c82c5e05c4f912a88eab0dcd91a171688f
 - name: github.com/spf13/cobra
-  version: 99b5d838ca16c25cc4944e323684f8415e8b10ba
+  version: 4d647c8944eb42504a714e57e97f244ed6344722
+  subpackages:
+  - doc
 - name: github.com/spf13/pflag
   version: e57e3eeb33f795204c1ca35f56c44f83227c6e66
 - name: github.com/timewasted/linode
