+++ hack/gendocs/make.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+pushd $GOPATH/src/github.com/appscode/voyager/hack/gendocs
+go run main.go
+
+cd $GOPATH/src/github.com/appscode/voyager/docs/reference
+sed -i 's/######\ Auto\ generated\ by.*//g' *
+popd
