+++ pkg/cmds/exporter_handler.go
@@ -1,4 +1,4 @@
-package main
+package cmds
 
 import (
 	"errors"
