+++ pkg/cmds/main.go
@@ -1,0 +1,32 @@
+package cmds
+
+import (
+	"flag"
+	"log"
+
+	v "github.com/appscode/go/version"
+	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
+	_ "k8s.io/client-go/kubernetes/fake"
+)
+
+func NewCmdVoyager(version string) *cobra.Command {
+	rootCmd := &cobra.Command{
+		Use:   "voyager [command]",
+		Short: `Voyager by Appscode - Secure Ingress Controller for Kubernetes`,
+		PersistentPreRun: func(c *cobra.Command, args []string) {
+			c.Flags().VisitAll(func(flag *pflag.Flag) {
+				log.Printf("FLAG: --%s=%q", flag.Name, flag.Value)
+			})
+		},
+	}
+	rootCmd.PersistentFlags().AddGoFlagSet(flag.CommandLine)
+	// ref: https://github.com/kubernetes/kubernetes/issues/17162#issuecomment-225596212
+	flag.CommandLine.Parse([]string{})
+
+	rootCmd.AddCommand(NewCmdRun(version))
+	rootCmd.AddCommand(NewCmdExport(version))
+	rootCmd.AddCommand(v.NewCmdVersion())
+
+	return rootCmd
+}
